clean up and tests
