+++ pIChemiSt/Makefile
@@ -1,5 +1,5 @@
 .PHONY: test
 
 test:
-	pytest -vvs test/test_cli.py
+	pytest -vvs test
 	# pytest -vs test/test_pka_sets.py
