+++ pIChemiSt/data/smarts/legacy/smarts_pKaMatcher.py
@@ -27,57 +27,3 @@ def read_smarts_pKaMatcher():
 
 # Read in the data
 list_smarts_pka = read_smarts_pKaMatcher()
-
-if __name__ == "__main__":
-    """
-    Conversion of DAT semi-structured data into JSON
-    """
-    import json
-    data = read_smarts_pKaMatcher()
-    print(data)
-    result = dict()
-    for e in data:
-        res_data = list()
-        for i in e:
-            smarts = i["smarts"]
-            name = i["name"]
-            res_data.append({
-                "pka": i["pka"],
-                "pka_std": i["pka_std"],
-                "idx": i["ind"],
-                "type": i["type"]
-            })
-
-            result[smarts] = {
-                "name": name,
-                "data": res_data
-            }
-    
-    # with open("smarts_converted.json", "w") as f:
-    #     json.dump({"smarts": result}, f, indent=4)
-
-    # create_logic_set_from_standardised_json
-    # print(result)
-    new_data = list()
-
-    for smarts, v in result.items():
-        res_data = list()
-        name = v["name"]
-        data = v["data"]
-        for d in data:
-            res_data.append({
-                "pka": d["pka"],
-                "ind": d["idx"],
-                "pka_std": d["pka_std"],
-                "type": d["type"],
-                "smarts": smarts,
-                "name": name
-            })
-        new_data.append(res_data)
-    
-    # assert new_data == data
-    # print(new_data)
-    with open("../expected.json") as f:
-        exp = json.load(f)
-    
-    assert exp == new_data
