+++ pIChemiSt/pichemist/pkamatcher.py
@@ -7,7 +7,11 @@ from pichemist.model import PKaType
 
 
 class PKaMatcher(object):
-    """TODO:..."""
+    """
+    Uses a set of SMARTS definitions and their pKa annotations
+    against input SMILES to match their pKa(s).
+
+    """
     def __init__(self):
         self.smarts_set = SS_SMARTS_PKA_SET
         self.skip_names = SKIP_SMARTS_NAMES
@@ -28,13 +32,16 @@ class PKaMatcher(object):
         # Each group is described as a list of dicts
         for pka_group_list in self.smarts_set:
             for pka_dict in pka_group_list:
+
                 # Skip if name is included in skip config
                 if pka_dict['name'] in self.skip_names:
                     continue
+
+                # Match SMARTS against the input and store pKa
                 used_idx_list = list()   # Used within larger scope
                 used_idx_local = list()  # Used within smaller scope
                 pat = Chem.MolFromSmarts(pka_dict['smarts'])
-                for match in mol.GetSubstructMatches(pat):      # ANDREY: There is no break here. Are we expecting multi matching?
+                for match in mol.GetSubstructMatches(pat):
                     match_idx = match[pka_dict['idx']-1]
                     used_idx_list += match
                     match_idxs = set(match)
@@ -44,14 +51,14 @@ class PKaMatcher(object):
                         raise PKaMatcherException(
                             f"Unknown site type (got {pka_dict['type']})")
 
-                    # Basic and acid pKa matching
-                    # TODO: See below
+                    # Basic
                     if pka_dict['type'] == PKaType.BASIC.value:
                         if match_idx in available_idxs \
                                 and match_idx not in used_idx_local:
                             if pka > PKA_LIMITS["base_1"] \
                                     and pka < PKA_LIMITS["base_2"]:
                                 results["base"].append((pka, smiles))
+                    # Acidic
                     if pka_dict['type'] == PKaType.ACIDIC.value:
                         if match_idx in available_idxs \
                                 and match_idx not in used_idx_local:
@@ -62,7 +69,14 @@ class PKaMatcher(object):
             all_used_idxs_set = all_used_idxs_set.union(set(used_idx_list))
         return results
 
+    def calculate_pka_from_smiles(self, smiles: str):
+        """Calculates the pKa values for a SMILES."""
+        results = self._pka_dict_from_smiles(smiles)
+        return {k: [t[0] for t in v]
+                for k, v in results.items()}
+
     def calculate_pka_from_list(self, smiles_list: list):
+        """Calculates the pKa values for a list of SMILES."""
         base_pkas = list()
         acid_pkas = list()
         diacid_pkas = list()    # ANDREY: not used
@@ -77,21 +91,3 @@ class PKaMatcher(object):
 class PKaMatcherException(Exception):
     def __init__(self):
         pass
-
-
-if __name__ == "__main__":
-    matcher = PKaMatcher()
-    unknown_fragments = ['CC(=O)N[C@@H](CCCN)C(C)=O',
-                         'CC(=O)N[C@@](C)(CC(=O)O)C(C)=O']
-    res = matcher.calculate_pka_from_list(unknown_fragments)
-    expected = ([(10.4, 'CC(=O)N[C@@H](CCCN)C(C)=O')],
-                [(3.46, 'CC(=O)N[C@@](C)(CC(=O)O)C(C)=O')], [])
-    assert res == expected, f"got {res}"
-
-    fragment = 'CC(=O)N[C@@H](CCCN)C(C)=O'
-    res = matcher._pka_dict_from_smiles(fragment)
-    # ANDREY: How possibly the same fragment can
-    # return both acid and basic or multiple values?
-    # This should be slimmed down into {"type": "acid/base", "value": 10.4}
-    expected = {'acid': [], 'base': [(10.4, 'CC(=O)N[C@@H](CCCN)C(C)=O')]}
-    assert res == expected, f"got {res}"
