+++ pIChemiSt/test/test_cli.py
@@ -1,13 +1,8 @@
 import os
-import json
 
 from pichemist import cli
-from pichemist.fasta.smarts import AA_SMARTS_SET
-from pichemist.fasta.pka_sets import FASTA_PKA_SETS
-from pichemist.smarts.pka_set import SS_SMARTS_PKA_SET
 
 script_dir = os.path.dirname(os.path.realpath(__file__))
-examples_dir = os.path.join(script_dir, "examples")
 
 
 def test_pka_matcher_json():
@@ -57,23 +52,4 @@ def test_pka_matcher_json():
                 }
 
     output = cli.calc_pIChemiSt(options)
-    # print(output)
     assert expected == output
-
-
-def test_fasta_pka_sets():
-    with open(f"{examples_dir}/fasta_pka_sets_expected.json") as f:
-        expected = json.load(f)
-    assert expected == FASTA_PKA_SETS
-
-
-def test_ss_smarts_pka_set():
-    with open(f"{examples_dir}/ss_smarts_pka_set_expected.json") as f:
-        expected = json.load(f)
-    assert expected == SS_SMARTS_PKA_SET
-
-
-def test_aa_smarts_pka_set():
-    with open(f"{examples_dir}/aa_smarts_pka_set_expected.json") as f:
-        expected = json.load(f)
-    assert expected == AA_SMARTS_SET
